sandbox: move ln-time symlink check into its own extension file
The symlink validation added in the previous commit pushed
MountedFileSystem.swift to 418 lines, tripping SwiftLint's default
file_length warning under --strict (the lint job pre-empts the rest
of CI so the symptom was a single annotation with no detail).

Split the helper into MountedFileSystem+Symlinks.swift — same pattern
as the existing MountedFileSystem+Synthesis.swift sibling — and drop
`resolve(_:)` from `private` to internal so the extension can reach
it. Main file is back to 394 lines.

No behaviour change.
